Create generic string vec fn
